Set doNotHandleEscape in InputController back to false, if it was set to true. Otherwise ExitMenuController.visible will not be set to true next time that the game opens and the user presses Escape.
Do not show/hide the view, unless it was/wasn't displayed previously. (Avoids excessive messages in console.)

Expose maximum timeout for awaits. Do not retrieve latch from Main and just store it locally.

Make Main count down a CountDownLatch when the game starts, if it was set.

Hide the exit menu when stopping the game.

Make sure that the back button has a mousePressed event handler.

Typo in method signature. (Accidentally included an inparameter for a simple getter.)

Restore getter for the inputController field and rename the getter for gameRunning to a more appropriate name.

Reset state everytime that the view is hidden.

Finish tests for main controllers.

Refactor the rebinding of controls, stop returning a boolean from rebindControl and make the keyPressed event work without a return value.

Check magazine counter vs. magazine size before reloading.

Expose getter for magazineSize.

Testing views is not required for this course.

Return ActionButtons from the button getters instead of the superclass (Node).

Move setting control button text to the controller instead, as the related code contains a bit of interaction with InputController.

Typo + expose getter for SettingsMenuController.mainMenuController.

Stop the game before shutting down.

Code cleanup. Allow Main to check whether or not tests are running. Important to avoid shutting down the FX instance as you are only allowed to launch the application once.

Refactoring.

Cleanup unused imports.

Automatically stop the game, if it was running when startGame was called.

Save the loaded level name and expose the getter for it.

Expose getter for mainMenuController